# backend/src/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 0.8.3
  title: ilSaye Backend API
  description: An event communication backend for large audiences written in Nodejs, with express and Sequelize
schemes:
  - https
  - http
servers:
  - url: https://ilsayebackend.azurewebsites.net/api/v1
    description: Main (production) server
  - url: http://localhost:8080/api/v1
    description: Internal staging server for testing
...
paths:
  /login:
    post:
      tags:
        - auth
      summary: standard user pass login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/authenticatedUserResponse'
        401:
          description: fail
        503:
          description: server error
  /logingoogle:
    post:
      tags:
        - auth
      summary: google oauth2
      description: validates an oauth token from user, making sure it is both valid and for the intended website
      parameters:
        - in: path
          name: token
          schema:
            type: string
          required: true
          description: token from google after successful login
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/authenticatedUserResponse'
        401:
          description: fail
        503:
          description: server error, response from google auth server wasn't understood
...
  /signup:
    post:
      tags:
        - auth
      summary: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type:
                object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: ok
        400:
          descritption: failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
...
components:
  schemas:
    authenticatedUserResponse:
      title: Authenticated user response
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        eventName:
          type: string
          nullable: true
          description: last joined event name
    errorResponse:
      title: Error Response
      type: object
      properties:
        errorCode:
          type: number
        msg:
          type: string
...